+++ src/main/java/kitchenpos/menuGroup/web/MenuGroupRestController.java
@@ -1,7 +1,7 @@
 package kitchenpos.menuGroup.web;
 
 import kitchenpos.menuGroup.model.MenuGroup;
-import kitchenpos.menuGroup.service.MenuGroupBo;
+import kitchenpos.menuGroup.application.MenuGroupBo;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
