+++ src/main/java/kitchenpos/product/web/ProductRestController.java
@@ -1,7 +1,7 @@
 package kitchenpos.product.web;
 
 import kitchenpos.product.model.Product;
-import kitchenpos.product.service.ProductBo;
+import kitchenpos.product.application.ProductBo;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
