+++ src/test/java/kitchenpos/bo/TableBoTest.java
@@ -5,7 +5,7 @@ import kitchenpos.table.model.OrderTableDao;
 import kitchenpos.order.model.Order;
 import kitchenpos.order.model.OrderStatus;
 import kitchenpos.table.model.OrderTable;
-import kitchenpos.table.service.TableBo;
+import kitchenpos.table.application.TableBo;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
